Part II - completed

+++ js/script-bw.js
@@ -19,8 +19,6 @@ function fetchPhotos() {
                 descriptionTitle.setAttribute('class', 'description-title');
                 descriptionTitle.innerText="Description";
 
-
-
                 let photo = document.createElement('img');
                 photo.setAttribute('class', 'photo');
                 photo.setAttribute('src', data[i].url);
@@ -37,18 +35,11 @@ function fetchPhotos() {
                     document.getElementById('description-input').value = data[i].description;
                     currentID = data[i].id;
                 }
-
-                let Delete = document.createElement('button');
-                Delete.innerText = 'Delete';
-                Delete.onclick = function() {
-                    deletePhoto(data[i].id);
-                }
                 
                 photoContainer.appendChild(photo);
                 photoContainer.appendChild(descriptionTitle);
                 photoContainer.appendChild(description);
                 photoContainer.appendChild(edit);
-                photoContainer.appendChild(Delete);
                 mainContainer.appendChild(photoContainer);
             }
         })
@@ -75,31 +66,40 @@ function add() {
 }
 
 function update() {
-    let url = document.getElementById('url').value;
-    let descriptionInput = document.getElementById('description-input').value;
-    
-    let newPhoto = {
-        url: url,
-        description: descriptionInput
+    if (currentID!==-1) {
+        let url = document.getElementById('url').value;
+        let descriptionInput = document.getElementById('description-input').value;
+        
+        let newPhoto = {
+            url: url,
+            description: descriptionInput
+        }
+
+        fetch('http://localhost:8080/photosBW/' + currentID, 
+        {
+            method: 'put',
+            headers:  {'Content-Type': 'application/json'},
+            body: JSON.stringify(newPhoto)
+        }).then(function(response) {
+            window.location.reload();
+        })
+
+        currentID = -1;
     }
 
-    fetch('http://localhost:8080/photosBW' + currentID, 
-    {
-        method: 'put',
-        headers:  {'Content-Type': 'application/json'},
-        body: JSON.stringify(newPhoto)
-    }).then(function(response) {
-        window.location.reload();
-    })
 }
 
-function deletePhoto(id) {
-    fetch('http://localhost:8080/photosBW' + id, 
-    {
-        method: 'delete'
-    }).then(function(response) {
-        window.location.reload();
-    })
+function deletePhoto() {
+    if (currentID!==-1)
+
+        fetch('http://localhost:8080/photosBW/' + currentID, 
+        {
+            method: 'delete'
+        }).then(function(response) {
+            window.location.reload();
+        })
+
+        currentID = -1;
 }
 
 fetchPhotos();
